GO-1421: read installation arguments from 'go env'
This change makes main read os/arch from the environment rather than
from the runtime.

Also:

    * cleans up logging so everything goes to stderr
    * corrects darwin/arm64 installs
    * tries to use the next best default download location if $GOBIN is
      not set

This also introduces filesystem-based tests using a harness which lets
us write scripts to validate behavior of the command. See 'go doc
testscript' for more details. This harness is almost certainly overkill
for our purposes and may be worth reconsidering if we are ok trading
extra boilerplate to get rid of the test dependencies.