pipelines: bump versions to go1.22 and go1.23 (#34)
This commit bumps the pipeline versions up to the two supported go
versions, go1.22 and go1.23.

There was one script test that intentionally created an invalid
GOPATH to ensure we handle that gracefully. Sine go1.22, go now
outputs a warning that the go.mod file is ignored instead of
breaking, which breaks this test. We already have another test case
for an invalid GOPATH, so I think we can safely remove this test.